#31: Moves Dash initialization into __main__.
